- refactoring code - add open in webstrom action
