+++ .gitignore
@@ -1,3 +1,4 @@
 info.plist
 persist
 node_modules
+config.js
