+++ action.js
@@ -1,0 +1,12 @@
+class Action {
+    constructor(options) {
+        this.actionName = options.actionName;
+        this.executor = options.executor;
+    }
+
+    execute(arg) {
+        this.executor(arg);
+    };
+}
+
+module.exports = Action;
