+++ config.sample.js
@@ -1,0 +1,10 @@
+{
+    "source-containers": [
+        "/Users/user_xxx/src"
+    ],
+
+    "sources": {
+    },
+
+    "stash-server": ""
+}
